move resources view
