+++ airflow/www/security.py
@@ -29,6 +29,7 @@ from airflow.auth.managers.fab.views.permissions import (
     ActionModelView,
     CustomRoleModelView,
     PermissionPairModelView,
+    ResourceModelView,
 )
 from airflow.auth.managers.fab.views.user import (
     CustomUserDBModelView,
@@ -50,9 +51,6 @@ from airflow.utils.log.logging_mixin import LoggingMixin
 from airflow.utils.session import NEW_SESSION, provide_session
 from airflow.www.extensions.init_auth_manager import get_auth_manager
 from airflow.www.fab_security.sqla.manager import SecurityManager
-from airflow.www.fab_security.views import (
-    ResourceModelView,
-)
 from airflow.www.utils import CustomSQLAInterface
 
 EXISTING_ROLES = {
